Move all non-main articles to articles/
The multi-column articles are too format-dependent to abstract away,
and are probably not going to be useful for handouts anyway: either a
handout will just be a section of a newspaper with some relevant
stories, or it'll be the full page where the main story is
plot-relevant.

Also add two Costa Rica filler articles: influenza-epidemic-over and
tinoco-statues-destroyed.